refactored main
